requirements.txt
open3d
torchmetrics
troch_geometric
pandas
pyyaml
wandb
matplotlib
pytorch_lightning